refactor(sync): move the registry reads out of the orchestrator
The five methods that read the local registry and the completion stamps —
the classify baseline, the per-unit bound-row projection, the
unstamped-platform count, the resident sibling-group keys and the
stale-row scan — move into services/library/registry_queries.py as
RegistryQueries. Their one dependency is the Unit-of-Work factory.

Behaviour is unchanged: the bodies moved verbatim, each opens the same
read Unit of Work at the same point, and each is still offloaded through
the orchestrator's executor with nothing open on the calling path. The
methods are peer-called now, so they lose their leading underscore and
take the do_ prefix the conventions give a peer-called worker.

Two neighbours deliberately stayed. The platform stamp's DELETE writes,
and its correctness is positional — after the fetch succeeded, after the
artwork, after the cancel guard, before the first chunk, so a crash in
between leaves no stamp (ADR-0023). It touches the same table through the
same repository as the unstamped-platform count and sits ten lines from
it: one answers a question the caller reasons about, the other is a move
the caller makes at a chosen moment. The component-key stamp does no I/O
at all — it calls a pure domain function and writes the result back onto
the caller's own list.

Counted size 1282 → 1180; the allowlist entry drops to match.
